Check C prototypes against Rust extern "C" signatures
Nothing verified that the ~61 prototypes in c_src/tim.h matched the Rust
functions they declare. The C compiler checks call sites against the header and
rustc checks the Rust function; neither compares the two, so a mismatch is
silent. This project already lost a session to one ABI bug of this family
(28c6ba5), and ~55 more functions are still to be ported the same way.

scripts/check-ffi-signatures.py compares every #[no_mangle] pub extern "C" fn
against its C prototype using the project's type mapping, and verify.sh now
fails the gate on a mismatch. It needs no game data, so it also runs on the
ALLOW_NO_GAME_DATA path.

It checks two independent things:

* ABI compatibility -- size class and pointer depth. This is the class that
  corrupts data.
* Signedness. u16 and s16 are ABI-identical so this is not an ABI bug, but
  widening sign-extends on one side and zero-extends on the other, and
  comparisons flip, so it is still a real defect.

Both were proven to fail by deliberately corrupting a prototype.

Mismatches found and fixed:

* part_explicit_size returned bool in C, c_int in Rust. Compatible only because
  this codebase does `typedef int bool`; declared int so it can be verified.
* part_mass was declared s16 in C but returned u16 from Rust, while the
  Part.mass field it feeds is i16. Rust now returns i16; every mass in the
  table is <= 28000 so the cast is the identity.

+++ src/tim_c.rs
@@ -1331,9 +1331,11 @@ pub extern "C" fn part_density(part_type: c_int) -> u16 {
 }
 
 #[no_mangle]
-pub extern "C" fn part_mass(part_type: c_int) -> u16 {
+pub extern "C" fn part_mass(part_type: c_int) -> i16 {
     let t = PartType::from_u16(part_type as u16);
-    parts::get_def(t).mass
+    // PartDef stores mass as u16; the C prototype and Part.mass are both s16. Every mass in
+    // the table is <= 28000, so this cast is the identity.
+    parts::get_def(t).mass as i16
 }
 
 #[no_mangle]
